# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Memoar API
  version: 0.1.1
  description: Contract for capture, archive, retrieval, sharing, conversion, and MCP authentication.
servers:
  - url: http://localhost:4000/v1
    description: Local development
...
paths:
...
  /sessions:
    get:
      tags: [Sessions]
      operationId: listSessions
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - { name: agent, in: query, schema: { type: string } }
        - { name: workspace, in: query, schema: { type: string } }
        - { name: machineId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: model, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Cursor page of session summaries
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionPage' }
...
  /sessions/{sessionId}:
    get:
      tags: [Sessions]
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - { name: cursor, in: query, schema: { type: string } }
        - { name: chunkSize, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Canonical session chunk
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionChunk' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
    patch:
      tags: [Sessions]
      operationId: updateSessionVisibility
      description: Changes a session's visibility scope. Widening beyond private requires a completed redaction review whose id is supplied as redactionReviewId and whose content digest still matches the session.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [visibility]
              properties:
                visibility:
                  type: object
                  required: [scope]
                  properties:
                    scope: { type: string, enum: [private, team, org, link] }
                    teamId: { $ref: '#/components/schemas/Uuid' }
                    orgId: { $ref: '#/components/schemas/Uuid' }
                redactionReviewId: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200':
          description: Updated visibility
          content:
            application/json:
              schema:
                type: object
                required: [id, visibility]
                properties:
                  id: { $ref: '#/components/schemas/Uuid' }
                  visibility: { $ref: '#/components/schemas/Visibility' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/RedactionReviewRequired' }
    delete:
      tags: [Sessions]
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '202': { description: Deletion and retention cleanup queued }
...
  /annotations:
    get:
      tags: [Annotations]
      operationId: listAnnotations
      parameters:
        - { name: sessionId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '200':
          description: Annotations visible in the active tenant, optionally filtered to one session
          content:
            application/json:
              schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Annotation' } } } }
    post:
      tags: [Annotations]
      operationId: createAnnotation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnnotationWrite' }
      responses:
        '201':
          description: Annotation created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Annotation' }
...
  /settings:
    get:
      tags: [Settings]
      operationId: getSettings
      responses:
        '200':
          description: Tenant settings with defaults applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantSettings' }
    put:
      tags: [Settings]
      operationId: updateSettings
      description: Partial update; omitted sections keep their current values. Custom redaction patterns must be valid regular expressions and a day-based retention policy requires days >= 1.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                redaction: { $ref: '#/components/schemas/RedactionSettings' }
                retention: { $ref: '#/components/schemas/RetentionSettings' }
      responses:
        '200':
          description: Updated tenant settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantSettings' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
  /distillation/settings:
    get:
      tags: [Distillation]
      operationId: getDistillationSettings
      responses:
        '200':
          description: Distillation opt-in and monthly budget state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DistillationSettings' }
    put:
      tags: [Distillation]
      operationId: updateDistillationSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: { type: boolean }
                provider: { type: string, enum: [none, anthropic] }
                model: { type: string, nullable: true }
                apiKey:
                  type: string
                  nullable: true
                  writeOnly: true
                  description: >-
                    Omit to leave the stored credential untouched; send null to
                    clear it. Never returned by any endpoint, and never written
                    to a log or an audit record.
                monthlyBudgetCents: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Updated distillation settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DistillationSettings' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
  /distillation/sessions/{sessionId}:
    post:
      tags: [Distillation]
      operationId: runDistillation
      description: Runs distillation for one session under the monthly cost cap and records a durable distill job. Requires prior opt-in via distillation settings.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '202':
          description: Distillation job completed or failed; inspect status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DistillationJob' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '409':
          description: Distillation is not opted in, or the estimated cost exceeds the remaining monthly budget
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /distillation/jobs/{jobId}:
    get:
      tags: [Distillation]
      operationId: getDistillationJob
      parameters:
        - { name: jobId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
      responses:
        '200':
          description: Durable distillation job record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DistillationJob' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
  /distillation/projects/export:
    post:
      tags: [Distillation]
      operationId: exportProjectMemory
      parameters:
        - { name: workspace, in: query, required: true, schema: { type: string } }
        - { name: format, in: query, schema: { type: string, enum: [claude, agents], default: agents } }
      responses:
        '200':
          description: Markdown project memory assembled from distilled notes
          content:
            application/json:
              schema:
                type: object
                required: [workspace, format, markdown, noteCount]
                properties:
                  workspace: { type: string }
                  format: { type: string, enum: [claude, agents] }
                  markdown: { type: string }
                  noteCount: { type: integer }
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKey: { type: apiKey, in: header, name: X-Memoar-Key }
    machineToken: { type: http, scheme: bearer, bearerFormat: MachineToken }
  parameters:
    SessionId: { name: sessionId, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
...
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
  responses:
    ProblemResponse:
      description: Request failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RedactionReviewRequired:
      description: A completed redaction review is required before visibility widens
      content:
        application/problem+json:
          schema: { allOf: [{ $ref: '#/components/schemas/Problem' }, { type: object, required: [reviewId], properties: { reviewId: { $ref: '#/components/schemas/Uuid' } } }] }
  schemas:
    Uuid: { type: string, format: uuid }
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail:
          type: string
          description: >
            Present on a 4xx, where the caller's own request was at fault and
            naming the problem helps them. Absent on a 5xx, whose cause belongs
            in the log rather than the response.
        requestId:
          type: string
          description: >
            Also returned as the x-request-id header and written on the server's
            log line for this request, so a report can be traced to it.
        retryable: { type: boolean }
...
    SessionSummary:
      type: object
      description: >
        Branch, machine, tokens and duration are here because a client that is
        not given them fills the gap: they were rendered as "unknown",
        "Archived machine", "0 tokens" and "0m" on every row. Each is optional
        and absent when the session genuinely does not have it.
      required: [id, title, source, sourceLabel, workspace, updatedAt, turnCount, tokenCount, redactionStatus]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        title: { type: string }
        summary: { type: string }
        source: { type: string }
        sourceLabel: { type: string, description: "The tool's own name, not a title-cased id." }
        machineId: { $ref: '#/components/schemas/Uuid' }
        workspace: { type: string }
        branch: { type: string }
        model: { type: string }
        updatedAt: { type: string, format: date-time }
        turnCount: { type: integer }
        tokenCount: { type: integer, minimum: 0 }
        durationMinutes: { type: integer, minimum: 0 }
        redactionStatus: { type: string, enum: [clear, findings, reviewed] }
        score: { type: number }
        highlight: { type: string }
    SessionPage:
      type: object
      required: [items, total, nextCursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/SessionSummary' } }
        total:
          type: integer
          minimum: 0
          description: >
            How many sessions match the filter, not how many this page holds.
            Without it a client can say how many sessions it is holding but not
            how many there are, which is the number a reader wants.
        nextCursor: { type: [string, 'null'] }
...
    SessionChunk:
      type: object
      required: [session, turns, provenance, nextCursor]
      properties:
        session: { $ref: '#/components/schemas/SessionSummary' }
        turns: { type: array, items: { type: object, additionalProperties: true } }
        provenance:
          type: array
          description: >-
            How this session reached the archive - native capture, import, conversion,
            or distillation. Carried on the detail chunk rather than the summary so
            list responses stay small.
          items: { $ref: '#/components/schemas/ProvenanceEntry' }
        nextCursor: { type: [string, 'null'] }
    ProvenanceEntry:
      type: object
      required: [kind, capturedAt]
      properties:
        kind: { type: string, enum: [native, import, conversion, distillation] }
        sourceId: { type: string }
        capturedAt: { type: string, format: date-time }
        parserVersion: { type: string }
        details: { type: object, additionalProperties: true }
...
    AnnotationWrite:
      type: object
      required: [sessionId, kind, value]
      properties:
        sessionId: { $ref: '#/components/schemas/Uuid' }
        turnId: { $ref: '#/components/schemas/Uuid' }
        blockId: { $ref: '#/components/schemas/Uuid' }
        kind: { type: string, enum: [tag, collection, pin, note, summary, redaction_mask] }
        value: { type: object, additionalProperties: true }
    Annotation:
      allOf:
        - { $ref: '#/components/schemas/AnnotationWrite' }
        - type: object
          required: [id, createdAt, updatedAt]
          properties:
            id: { $ref: '#/components/schemas/Uuid' }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    RedactionSettings:
      type: object
      required: [secretScan, pathScan, emailScan, customPatterns]
      properties:
        secretScan: { type: boolean }
        pathScan: { type: boolean }
        emailScan: { type: boolean }
        customPatterns: { type: array, items: { type: string }, maxItems: 64 }
    RetentionSettings:
      type: object
      required: [policy, exemptCollected]
      properties:
        policy: { type: string, enum: [indefinite, days] }
        days: { type: integer, minimum: 1 }
        exemptCollected: { type: boolean }
    DistillationSettings:
      type: object
      description: >-
        Distillation is the only feature that sends archived content to a third
        party, so the account that owns the sessions brings its own provider
        credential and pays for its own use. The credential itself is never
        returned by any endpoint: keySet says whether one is stored and keyHint
        is its last four characters, which is enough to recognise which key it
        is and not enough to use it.
      required: [enabled, provider, keySet, monthlyBudgetCents, monthlySpentCents, remainingCents, budgetWindowStartedAt]
      properties:
        enabled: { type: boolean }
        provider:
          type: string
          enum: [none, anthropic]
          description: An enum rather than a boolean, so adding a provider is a new value and not a second field.
        model: { type: string, nullable: true }
        keySet: { type: boolean }
        keyHint: { type: string, nullable: true, description: Last four characters of the stored credential. }
        monthlyBudgetCents: { type: integer, minimum: 0 }
        monthlySpentCents: { type: integer, minimum: 0 }
        remainingCents: { type: integer, minimum: 0 }
        budgetWindowStartedAt: { type: string, format: date-time }
    DistillationJob:
      type: object
      required: [id, sessionId, status, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        sessionId: { $ref: '#/components/schemas/Uuid' }
        status: { type: string, enum: [queued, running, ready, failed] }
        createdAt: { type: string, format: date-time }
        noteIds: { type: array, items: { $ref: '#/components/schemas/Uuid' } }
        costCents: { type: integer }
        inputTokens: { type: integer }
        outputTokens: { type: integer }
        remainingCents: { type: integer }
        error: { type: string }
    TenantSettings:
      type: object
      required: [redaction, retention]
      properties:
        redaction: { $ref: '#/components/schemas/RedactionSettings' }
        retention: { $ref: '#/components/schemas/RetentionSettings' }
        updatedAt: { type: [string, 'null'], format: date-time }
...